Refactor combat-distances into smaller parts
